working version of material css cards with plotly on reveal served via ajax
hamster wheeling displayed until loaded. lots of debut alerts to be removed. handle_plot_format still needs more attention to make this work outside of localhost:5000 and in a module. lib directory needs to be properly controlled and _documented_, the replacement string for the lib src in the widget must become an argument to reflect the actual deployment.